Remove string2char helper
Unless I'm missing something I believe this is ill-formed? The String is passed by copy and this returns a pointer to the underlying heap memory, so it's possible that the memory will get overwritten before the string is used.

Instead we can just pass the c_str() directly, as the function doesn't care about const-ness. This eliminates the compiler warning that function may fail to return a pointer if the string is 0 length, and avoids potentially passing nullptr to the getTextBounds function.